fix(ui): move Run now + Queue to Schedules tab and auto-dismiss toast (v2.4.13)
Two operator-dashboard UX fixes bundled into v2.4.13.

1. Jobs tab cleanup. The Run now button strip and Queue and schedule
   table moved off the Jobs tab to a dedicated Schedules tab inserted
   between Jobs and Files. New tab order: Jobs | Schedules | Files |
   Configuration. Jobs tab now only has search, type filter, refresh,
   and the recent-runs table. All existing behavior (manual triggering,
   1s burst polling on click, cron and last-run display) is preserved
   in the new tab.

2. Auto-dismissing toast. The Run now success/warning/error toast no
   longer sits on screen forever. Each toast auto-dismisses 4 seconds
   after it was raised, has its own timer (a fresh trigger does not
   extend an older toast), and now also has a manual X close button.
   The dismiss timer is cleared on tab unmount to avoid leaks.

Frontend test added: SchedulesTab.test.tsx asserts the toast is
removed from the DOM after ALERT_AUTO_DISMISS_MS. Vitest + jsdom +
@testing-library/react added as devDependencies along with an
'npm test' script.
